samples: nrf9160: mmss: Split connection.c into two files
Split `connection.c` into two files better representing functional
separation: `message_queue.c` for the message queue, and
`cloud_connection.c` for the connection loop.
