+++ samples/cellular/nrf_cloud_mqtt_multi_service/CMakeLists.txt
@@ -13,7 +13,8 @@ zephyr_compile_definitions(PROJECT_NAME=${PROJECT_NAME})
 # NORDIC SDK APP START
 target_sources(app PRIVATE src/main.c)
 target_sources(app PRIVATE src/application.c)
-target_sources(app PRIVATE src/connection.c)
+target_sources(app PRIVATE src/cloud_connection.c)
+target_sources(app PRIVATE src/message_queue.c)
 target_sources(app PRIVATE src/temperature.c)
 target_sources(app PRIVATE src/fota_support.c)
 target_sources(app PRIVATE src/led_control.c)
