+++ samples/cellular/nrf_cloud_mqtt_multi_service/src/application.h
@@ -1,4 +1,4 @@
-/* Copyright (c) 2022 Nordic Semiconductor ASA
+/* Copyright (c) 2023 Nordic Semiconductor ASA
  *
  * SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
  */
